job_automation/requirements.txt
playwright>=1.40.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
requests>=2.31.0
google-generativeai>=0.3.0
openai>=1.0.0
jinja2>=3.1.0
